requirements.txt
geopy
streamlit
python-dotenv
portalocker
plotly
dash
pycountry
vega-datasets
altair
gpdvega
